chore: bump mocha to version 10.0.0
Related: https://github.com/socketio/socket.io/issues/3710
